July 20th:
	- Complete the metrics on Statistics.

+++ Simulator/src/experiments/testbeds/DataCenter.java
@@ -5,7 +5,7 @@ import system.utility.Statistics;
 
 public class DataCenter extends Testbed {
 
-	public DataCenter(int networkType) {
+	public DataCenter(short networkType) {
 		super(networkType);
 	}
